Added ifdefs for ESP BSP capabilities.
This allows us to build on systems that don't have displays/audio/touch/buttons etc.  Also, git decided that *this* is the time to change all the files from CRLF to LF, and so changed all the files.  No idea why it didn't before, but now this commit is unreadable.  Have fun.